Update proj_try1.py
The list with index and  the list with actual elements are now in the right positions when calling the plot function. They were initially in opposite position, meaning the x parameter in the plot function was taking the value of elements and the height parameter in the bar() was taking the value of index. This was supposed to be opposite. This commit fixes this issue.