Quick Sort
This update contains the visualisation of the quick-sort sorting algorithm. For visualisation I have used matplotlib. The idea for displaying the data is still by using naive approach. Meaning I am still not using the animation method and therefore to update the graph, the graph is cleared and all the elements are re-plotted again.